todo: refactor all code to use EditorBuffer
