+++ src/focus.rs
@@ -40,13 +40,12 @@ pub(crate) fn add_editor_to_focused(
     q: Query<&CosmicEditBuffer, Without<CosmicEditor>>,
 ) {
     if let Some(e) = active_editor.0 {
-        let Ok(CosmicEditBuffer(b)) = q.get(e) else {
+        let Ok(buffer) = q.get(e) else {
             return;
         };
-        let mut editor = Editor::new(b.clone());
-        editor.set_redraw(true);
+        let editor = CosmicEditor::clone_from_buffer(buffer);
         trace!("Adding editor to focused widget");
-        commands.entity(e).insert(CosmicEditor::new(editor));
+        commands.entity(e).insert(editor);
     }
 }
 
@@ -56,20 +55,23 @@ pub(crate) fn drop_editor_unfocused(
     active_editor: Res<FocusedWidget>,
     mut q: Query<(Entity, &mut CosmicEditBuffer, &CosmicEditor)>,
 ) {
-    if active_editor.0.is_none() {
-        for (e, mut b, editor) in q.iter_mut() {
-            b.lines = editor.with_buffer(|buf| buf.lines.clone());
-            b.set_redraw(true);
-            trace!("Removing editor from all entities as there is no focussed widget",);
-            commands.entity(e).remove::<CosmicEditor>();
-        }
-    } else if let Some(focused) = active_editor.0 {
-        for (e, mut b, editor) in q.iter_mut() {
-            if e != focused {
-                b.lines = editor.with_buffer(|buf| buf.lines.clone());
-                b.set_redraw(true);
-                trace!("Removing editor from entity as its not focussed anymore",);
+    match active_editor.0 {
+        None => {
+            for (e, mut buffer, editor) in q.iter_mut() {
+                // buffer.lines = editor.with_buffer(|buf| buf.lines.clone());
+                // buffer.set_redraw(true);
+                *buffer = CosmicEditBuffer::from_downgrading_editor(&editor);
+                trace!("Removing editor from all entities as there is no focussed widget",);
                 commands.entity(e).remove::<CosmicEditor>();
+            }
+        }
+        Some(focused) => {
+            for (e, mut b, editor) in q.iter_mut() {
+                if e != focused {
+                    *b = CosmicEditBuffer::from_downgrading_editor(&editor);
+                    trace!("Removing editor from entity as its not focussed anymore",);
+                    commands.entity(e).remove::<CosmicEditor>();
+                }
             }
         }
     }
